requirements.txt
numPy
matplotlib
pillow


